op type extract
