+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/OpsPage/opsPageTypes.ts
@@ -1,0 +1,3 @@
+export type WFHighLevelOpVersionFilter = {
+  opName?: string | null;
+};
